Added code that can handle the removal of clients, all children of a client are moved when the client is removed, some adjustments were needed such as code where a list of clients is created but it seems to work well as of now. There is a problem with the clients page where the list does not refresh correctly when a client has just been deleted.
There also needs to be an investigation into how to deal with money sources since as of now the expectation is that the source will simply be the entity ID. But with clients being deleted this may be a problem. It may need to simply switch to a string of the name of the source, or clients may not be simply deleted but marked as deleted.
I'll need to consider this later.